+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 * Use `ViewModelMapping` model from `DTModelStorage`. Use typecasting instead of runtime introspection to map models.
 * Replace `CollectionViewReaction` class with `UIReaction` class from `DTModelStorage`
 * Allow user to customize view model mappings via `DTViewModelMappingCustomizable` protocol.
+* Better error handling via `DTCollectionViewFactory` error types. 
+* Allow custom closure to be run before fatal error in collection view factory will crash the application.
 
 ## [4.3.0](https://github.com/DenHeadless/DTCollectionViewManager/releases/tag/4.3.0)